Login Logout
Logout page has some buttons on it. Icon Appears on each page in the nav bar, and both login and logout links are displayed in the drop down. The buttons are now round and most of the terrible style tags have been relocated to styling from the style.css file